Move BinaryReaderWrapper to the correct namespace
