+++ SAGESharp.Tests/IO/BinarySerializerTests.cs
@@ -8,6 +8,7 @@ using NSubstitute;
 using NSubstitute.ClearExtensions;
 using NUnit.Framework;
 using SAGESharp.IO;
+using SAGESharp.IO.Binary;
 using System;
 using System.IO;
 using System.Reflection;
